// apps/web/lib/api/generated/orval/model/tourVariantResponseDto.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { TourSummaryDto } from './tourSummaryDto';

export type TourVariantResponseDto = {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  id: number | string;
  tour: TourSummaryDto;
  name: string;
  description: string;
  isPrimary: boolean;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  totalDistanceMeters: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  totalDurationMinutes: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  stageCount: number | string;
  createdAt: string;
};